+++ sr/models/__init__.py
@@ -6,6 +6,7 @@ from sr.models.band_adapter import BandAdapter
 from sr.models.band_reference import BandReference
 from sr.models.base import Base
 from sr.models.generation_job import GenerationJob
+from sr.models.instrument_slot import InstrumentSlot
 from sr.models.player import Player
 from sr.models.project import Project
 from sr.models.render_cache import RenderCache
@@ -23,6 +24,7 @@ __all__ = [
     "BandAdapter",
     "Project",
     "Player",
+    "InstrumentSlot",
     "Singer",
     "Song",
     "SongSection",